requirements.txt
asgiref==3.2.6
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
Markdown==3.2.1
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
